Project name tagging
Adds support for setting a project name in one or more documents
via the integration plugin. Project names are set as special-purpose
tags in the client, with a tag type of 10000. Tagging is performed
automatically whenever the citations in the document are updated.
Updates to tagging are cumulative, to allow multiple documents to
share the same project name.

In the client, project names behave as follows:

* In the tag selector:
    - Project names are displayed immediately
      after colored tags, with distinctive styling.
    - If both project name and an ordinary tag of the same name
      exist, the two are displayed separately.
    - Project names can be deleted in the tag selector; other
      operations (renaming, colorizing) are not available.
    - Deleting a project name or ordinary tag affects only
      the name/tag of that type.
    - Adding or renaming an ordinary tag with the same name
      as a project name has no effect on the project name or
      its dependent items.
    - HOWEVER: selecting a project name will also select
      an ordinary tag of the same time, if one exists.

* In the item tag box:
    - Project names do no appear in the tags box of individual items.